# Conflicts:
#	arius-admin/arius-admin-biz/src/main/java/com/didichuxing/datachannel/arius/admin/biz/cluster/impl/ClusterLogicManagerImpl.java
#	arius-admin/arius-admin-core/src/main/java/com/didichuxing/datachannel/arius/admin/core/service/cluster/logic/ClusterLogicService.java
#	arius-admin/arius-admin-core/src/main/java/com/didichuxing/datachannel/arius/admin/core/service/cluster/logic/impl/ClusterLogicServiceImpl.java

+++ arius-admin/arius-admin-core/src/main/java/com/didichuxing/datachannel/arius/admin/core/service/cluster/logic/impl/ClusterLogicServiceImpl.java
@@ -42,6 +42,14 @@ import com.didiglobal.logi.log.LogFactory;
 import com.didiglobal.logi.security.common.vo.project.ProjectVO;
 import com.didiglobal.logi.security.service.ProjectService;
 import com.google.common.collect.Lists;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -130,11 +138,12 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
      *
      * @param logicClusterId 资源id
      * @param operator       操作人
+     * @param projectId
      * @return result
      */
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public Result<Void> deleteClusterLogicById(Long logicClusterId, String operator) throws AdminOperateException {
+    public Result<Void> deleteClusterLogicById(Long logicClusterId, String operator, Integer projectId) throws AdminOperateException {
         ClusterLogicPO logicCluster = logicClusterDAO.getById(logicClusterId);
         if (logicCluster == null) {
             return Result.buildNotExist("逻辑集群不存在");
@@ -142,6 +151,11 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
 
         if (hasLogicClusterWithTemplates(logicClusterId)) {
             return Result.build(ResultType.IN_USE_ERROR.getCode(), "逻辑集群使用中");
+        }
+        final Result<Void> result = ProjectUtils.checkProjectCorrectly(ClusterLogicPO::getProjectId, logicCluster,
+                projectId);
+        if (result.failed()){
+            return result;
         }
 
         boolean succeed = (logicClusterDAO.delete(logicClusterId) > 0);
@@ -169,7 +183,7 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
      */
     @Override
     public Result<Long> createClusterLogic(ESLogicClusterDTO param) {
-        Result<Void> checkResult = validateClusterLogicParams(param, ADD);
+        Result<Void> checkResult = validateClusterLogicParams(param, ADD, param.getProjectId());
         if (checkResult.failed()) {
             LOGGER.warn("class=ClusterLogicServiceImpl||method=createClusterLogic||msg={}", checkResult.getMessage());
             return Result.buildFrom(checkResult);
@@ -187,16 +201,17 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
      *
      * @param param     参数
      * @param operation 操作
+     * @param projectId
      * @return result
      */
     @Override
-    public Result<Void> validateClusterLogicParams(ESLogicClusterDTO param, OperationEnum operation) {
-        return checkLogicClusterParams(param, operation);
+    public Result<Void> validateClusterLogicParams(ESLogicClusterDTO param, OperationEnum operation, Integer projectId) {
+        return checkLogicClusterParams(param, operation,projectId);
     }
 
     @Override
-    public Result<Void> editClusterLogic(ESLogicClusterDTO param, String operator) {
-        Result<Void> checkResult = validateClusterLogicParams(param, EDIT);
+    public Result<Void> editClusterLogic(ESLogicClusterDTO param, String operator, Integer projectId) {
+        Result<Void> checkResult = validateClusterLogicParams(param, EDIT,projectId);
         if (checkResult.failed()) {
             LOGGER.warn("class=ClusterLogicServiceImpl||method=editResource||msg={}", checkResult.getMessage());
             return checkResult;
@@ -441,6 +456,15 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
         return new ArrayList<>(pluginMap.values());
     }
 
+    /**
+     * @param clusterLogicId 集群逻辑id
+     * @return
+     */
+    @Override
+    public Integer getProjectIdById(Long clusterLogicId) {
+        return logicClusterDAO.getProjectIdById(clusterLogicId);
+    }
+
     @Override
     public Result<Long> addPlugin(Long logicClusterId, PluginDTO pluginDTO, String operator) {
 
@@ -494,15 +518,26 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
         return ConvertUtil.list2List(logicClusterDAO.listByIds(new HashSet<>(clusterLogicIdList)), ClusterLogic.class);
     }
 
+    @Override
+    public ClusterLogicDiskUsedInfoPO getDiskInfo(Long id) {
+        ClusterRegion clusterRegion =clusterRegionService.getRegionByLogicClusterId(id);
+        List<ESClusterRoleHostPO> esClusterRoleHostPOS = clusterRoleHostDAO.listByRegionId(Math.toIntExact(clusterRegion.getId()));
+        //节点名称列表
+        List<String> nodeList = esClusterRoleHostPOS.stream().map(ESClusterRoleHostPO::getNodeSet).collect(toList());
+        String clusterName =clusterRegion.getPhyClusterName();
+        return ariusStatsNodeInfoESDAO.getClusterLogicDiskUsedInfo(clusterName, nodeList);
+    }
+
     /***************************************** private method ****************************************************/
     /**
      * Check逻辑集群参数
      *
      * @param param     逻辑集群
      * @param operation 操作类型
+     * @param projectId
      * @return
      */
-    private Result<Void> checkLogicClusterParams(ESLogicClusterDTO param, OperationEnum operation) {
+    private Result<Void> checkLogicClusterParams(ESLogicClusterDTO param, OperationEnum operation, Integer projectId) {
         if (AriusObjUtils.isNull(param)) {
             return Result.buildParamIllegal("逻辑集群信息为空");
         }
@@ -530,6 +565,22 @@ public class ClusterLogicServiceImpl implements ClusterLogicService {
             ClusterLogicPO oldPO = logicClusterDAO.getById(param.getId());
             if (oldPO == null) {
                 return Result.buildNotExist("逻辑集群不存在");
+            }
+            //当param中projectid存在
+            if (Objects.nonNull(param.getProjectId())) {
+                final Result<Void> result = ProjectUtils.checkProjectCorrectly(ClusterLogicPO::getProjectId, oldPO,
+                        param.getProjectId());
+                if (result.failed()) {
+                    return result;
+                }
+
+            } else {
+                //校验路径
+                final Result<Void> result = ProjectUtils.checkProjectCorrectly(ClusterLogicPO::getProjectId, oldPO,
+                       projectId);
+                if (result.failed()) {
+                    return result;
+                }
             }
         }
